+++ CHANGES.md
@@ -1,4 +1,11 @@
 
+## 0.3
+
+Remove the ``BaseView`` class and the method selecting functionality, as it
+just doesn't work with Pyramid's view configuration.
+
+Re-factored the CSRF validation into a subscriber as a consequence.
+
 ## 0.2.1
 
 Fixed the `.i18n.add_underscore_translation` subscriber.